+++ PostgreEx.Infrastructure/Postgres/NpgsqlCostSettingsProvider.cs
@@ -5,7 +5,7 @@ internal class NpgsqlCostSettingsProvider(NpgsqlQueryExecutor queryExecutor)
     public async Task<CostSettings> GetCostSettings()
     {
         var (seqPageCost, cpuTupleCost) = (await queryExecutor
-                .Execute<decimal, decimal>("select current_setting('seq_page_cost'), current_setting('cpu_tuple_cost')")
+                .Execute<double, double>("select current_setting('seq_page_cost'), current_setting('cpu_tuple_cost')")
             )!;
 
         return new CostSettings(
